Added role validation
